swaylock: add additional aliases to my home monitor config
Essentially, I can't seem to have a consistent device associated to my displays
when I close/reopen them when connected to my dock.

AFAIK there is no way to parameterize this with `udev` and `swaylock` does not
allow you to use the "monitor name" like `sway` for its `output` statement.