waybar: copy someone else's config
The Spotify script is my own!

I tried using playerctl's cli tool and monitoring to continuously emit
data like my current script does but it didn't work.

My biggest motivation for the script is that polling playerctl with a
1 second interval was "too slow". I want to see my taskbar updated
immediately when I change a song, play, pause or disconnect Spotify.

Source for the style:
- <https://old.reddit.com/r/swaywm/comments/kf7zcs/frosty/>
- <https://github.com/nboughton/dotfiles/blob/d93a6b0564ebb6d0ea0c671e60838b969422174d/waybar/config>