Add pipewire profile under 'core-linux'
Pipewire is a replacement for pulseaudio and JACK though it integrates
and works with both protocols.

Pipewire is already in use for screen sharing under Wayland.

I've been having lots of problem with bluez and pulseaudio with my
bluetooth headset so I decided to use pipewire instead.

My bluetooth headset would connect but would not be registered as a
pulseaudio sink even though `bluetoothctl info` would show the device
as having the required capabilities.

With Pipewire, it worked without any issue or configuration.